refactor(terraform): move action handlers to `handlers.ts`
Like it's done in the terraform plugin.
